# Optimize MapVector::compare