Summary:
MapVector::compare sorts map entries by key using std::sort and
BaseVector::compare:

```
    std::sort(
        indices.begin(),
        indices.end(),
        [&](vector_size_t left, vector_size_t right) {
          return keys_->compare(keys_.get(), left, right) < 0;
        });
```

This is very slow. An optimization is to add BaseVector::sortIndices method and
specialize it for flat vectors so that type-dispatch happens only once per map.

Before:

```
============================================================================
velox/benchmarks/basic/VectorCompare.cpp     relative  time/iter   iters/s
============================================================================
compareSimilarSimpleFlat                                    6.58ns   151.94M
compareSimilarSimpleFlatNoDispatch                          2.31ns   432.64M
compareSimilarArray                                        90.45ns    11.06M
CompareSimilarMap                                           1.06us   940.21K
CompareSimilarRow                                          64.53ns    15.50M
----------------------------------------------------------------------------
```

After:

```
============================================================================
velox/benchmarks/basic/VectorCompare.cpp     relative  time/iter   iters/s
============================================================================
compareSimilarSimpleFlat                                    6.58ns   151.92M
compareSimilarSimpleFlatNoDispatch                          2.40ns   417.03M
compareSimilarArray                                        98.63ns    10.14M
compareSimilarMap                                         635.94ns     1.57M
compareSimilarRow                                          53.80ns    18.59M
----------------------------------------------------------------------------
```

Differential Revision: D40035913